requirements.txt
openai>=1
python-dotenv
